+++ substratevm/src/com.oracle.svm.core.windows/src/com/oracle/svm/core/windows/WindowsFeature.java
@@ -26,13 +26,19 @@ package com.oracle.svm.core.windows;
 
 import org.graalvm.nativeimage.ImageSingletons;
 
-import com.oracle.svm.shared.feature.AutomaticallyRegisteredFeature;
 import com.oracle.svm.core.feature.InternalFeature;
+import com.oracle.svm.core.imagelayer.ImageLayerBuildingSupport;
 import com.oracle.svm.core.log.Log;
 import com.oracle.svm.core.os.ImageHeapProvider;
+import com.oracle.svm.shared.feature.AutomaticallyRegisteredFeature;
 
 @AutomaticallyRegisteredFeature
 class WindowsFeature implements InternalFeature {
+
+    @Override
+    public boolean isInConfiguration(IsInConfigurationAccess access) {
+        return ImageLayerBuildingSupport.firstImageBuild();
+    }
 
     @Override
     public void duringSetup(DuringSetupAccess access) {